requirements.txt
flask==2.0.1
werkzeug==2.0.3
opencv-python>=4.8.0
numpy>=1.21.2
python-dotenv==1.0.0
